+++ sismic/bdd/__init__.py
@@ -1,4 +1,4 @@
-from .cli import execute_cli
-from .helpers import map_action, map_assertion
+from .bdd import map_action, map_assertion, execute_bdd
 
-__all__ = ['execute_cli', 'map_action', 'map_assertion']
+__all__ = ['execute_bdd', 'map_action', 'map_assertion']
+